backend/requirements.txt
# 웹 프레임워크
flask==3.0.0
flask-cors==4.0.0
flask-socketio>=5.3.0
pydantic>=2.10.0

# 데이터베이스
pymongo>=4.6.0

# 데이터 처리
pandas==2.1.4
numpy==1.26.2
ta>=0.10.0

# 거래소 API
websockets==12.0
python-binance==1.0.19
ccxt==4.1.87

# 비동기 처리
aiohttp==3.9.1
pytest-asyncio==0.23.2

# 설정 및 유틸리티
pyyaml>=6.0.0
python-dotenv==1.0.0
rich>=13.0.0

# 보안
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.0

# 테스트
pytest==7.4.3
